+++ spec/controllers/users_controller_spec.rb
@@ -15,7 +15,7 @@ RSpec.describe UsersController, type: :controller do
 
       expect(response.status).to eq(200)
       expect(controller.instance_variable_get(:@user).name).to eq('Different')
-      expect(controller.instance_variable_get(:@user).updater).to eq(@hera)
+      expect(controller.instance_variable_get(:@user).updater_id).to eq(@hera.id)
     end
   end
 
@@ -25,7 +25,7 @@ RSpec.describe UsersController, type: :controller do
       request.session = { user_id: @zeus.id }
 
       post :update, id: @hera.id, user: { name: 'Different Second' }
-      expect(controller.instance_variable_get(:@user).updater).to eq(@zeus)
+      expect(controller.instance_variable_get(:@user).updater_id).to eq(@zeus.id)
     ensure
       request.session = old_request_session
     end