+++ spec/dummy/app/models/person.rb
@@ -1,4 +1,3 @@
 class Person < ActiveRecord::Base
-  stampable
   model_stamper
 end