+++ spec/dummy/app/models/user.rb
@@ -1,4 +1,3 @@
 class User < ActiveRecord::Base
-  stampable
   model_stamper
 end